Adopt shopify_function_target to fix deprecations
