+++ order-routing/rust/local-pickup-delivery-option-generators/default/src/main.rs
@@ -1,33 +1,7 @@
-use shopify_function::prelude::*;
-use shopify_function::Result;
+use std::process;
+pub mod run;
 
-use serde::{Deserialize, Serialize};
-
-generate_types!(
-    query_path = "./input.graphql",
-    schema_path = "./schema.graphql"
-);
-
-#[derive(Serialize, Deserialize, Default, PartialEq)]
-struct Config {}
-
-#[shopify_function]
-fn function(_input: input::ResponseData) -> Result<output::FunctionResult> {
-    let operations = vec![output::Operation {
-        add: output::LocalPickupDeliveryOption {
-            title: Some("Main St.".to_string()),
-            cost: Some(Decimal(1.99)),
-            pickup_location: output::PickupLocation {
-                location_handle: "2578303".to_string(),
-                pickup_instruction: Some("Usually ready in 24 hours.".to_string()),
-            },
-        },
-    }];
-
-    // Build operations based on the input query response here.
-
-    Ok(output::FunctionResult { operations })
+fn main() {
+    eprintln!("Please invoke a named export.");
+    process::exit(1);
 }
-
-#[cfg(test)]
-mod tests;
